requirements.txt
django
gunicorn
faker
hurry.filesize